+++ django/applications/catmaid/urls.py
@@ -415,6 +415,8 @@ urlpatterns += [
     url(r'^(?P<project_id>\d+)/similarity/queries/(?P<similarity_id>\d+)/recompute$', similarity.recompute_similarity),
     url(r'^(?P<project_id>\d+)/similarity/queries/(?P<similarity_id>\d+)/storage$', similarity.SimilarityStorageDetail.as_view()),
     url(r'^(?P<project_id>\d+)/similarity/test-setup$', similarity.test_setup),
+    url(r'^(?P<project_id>\d+)/similarity/(?P<similarity_id>\d+)/cluster$', similarity.SimilarityClusterDetail.as_view()),
+    url(r'^(?P<project_id>\d+)/similarity/cluster/enabled$', similarity.SimilarityClusterMgmt.as_view()),
 ]
 
 # Cropping